+++ src/main/java/com/example/demo/core/repository/BroadcastRepository.java
@@ -1,12 +1,14 @@
 package com.example.demo.core.repository;
 
-import com.example.demo.core.entity.Broadcast;
 import java.util.List;
+import java.math.BigInteger;
 
 import org.springframework.stereotype.Repository;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
+import com.example.demo.core.entity.Broadcast;
+
 @Repository
-public interface BroadcastRepository extends MongoRepository<Broadcast, Long> {
+public interface BroadcastRepository extends MongoRepository<Broadcast, BigInteger> {
   List<Broadcast> findAll();
 }
